refactor_adminMemberPage
